requirements.txt
anthropic==0.34.2
streamlit==1.36.0
Pillow
opencv-python==4.8.0.74
git+https://github.com/THU-MIG/yolov10.git
huggingface-hub
dill==0.3.8
openai==1.34.0
langchain
python-dotenv
altair<5
transformers==4.38.2
numpy
zipfile36
pdf2image
pytest
python-docx
PyMuPDF
pytest
PyPDF2